drivers: mmidp: Set correct data type for aux counter
The register can take u32 value. So modify the function prototype to
accept u32 data type limit instead of u16. This helps the aux part
track correct duration of time.


Acked-for-series: Varunkumar Allagadapa <varunkumar.allagadapa@amd.com>

+++ XilinxProcessorIPLib/drivers/mmidp/src/xmmidp.h
@@ -467,7 +467,7 @@ void XMmiDp_SetPhyPowerdown(XMmiDp *InstancePtr, XMmiDp_PhyPwrDown Control);
 void XMmiDp_SetPhyWidth(XMmiDp *InstancePtr, XMmiDp_PhyWidth Width);
 void XMmiDp_SetAux_250Us_Cnt_Limit(XMmiDp *InstancePtr, u16 Limit);
 void XMmiDp_SetAux_2000Us_Cnt_Limit(XMmiDp *InstancePtr, u16 Limit);
-void XMmiDp_SetAux_100000Us_Cnt_Limit(XMmiDp *InstancePtr, u16 Limit);
+void XMmiDp_SetAux_100000Us_Cnt_Limit(XMmiDp *InstancePtr, u32 Limit);
 void XMmiDp_SetPmConfig1(XMmiDp *InstancePtr, u32 Val);
 void XMmiDp_SetPmConfig2(XMmiDp *InstancePtr, u32 Val);
